🌍 #32 Fix sampling of pure properties, delete meaningless tests about arity
