+++ tests/GalaxyCheck.Tests.V2/GalaxyCheck.Tests.V2.csproj
@@ -12,8 +12,8 @@
     <PackageReference Include="FluentAssertions" Version="5.10.3" />
     <PackageReference Include="Microsoft.NET.Test.Sdk" Version="16.7.1" />
     <PackageReference Include="Moq" Version="4.16.1" />
-    <PackageReference Include="NebulaCheck" Version="0.0.0-699119383" />
-    <PackageReference Include="NebulaCheck.Xunit" Version="0.0.0-699119383" />
+    <PackageReference Include="NebulaCheck" Version="0.0.0-763023610" />
+    <PackageReference Include="NebulaCheck.Xunit" Version="0.0.0-763023610" />
     <PackageReference Include="Snapshooter.Xunit" Version="0.5.8" />
     <PackageReference Include="xunit" Version="2.4.1" />
     <PackageReference Include="xunit.runner.visualstudio" Version="2.4.3">
