+++ tests/GalaxyCheck.Tests.V2/GenTests/InfiniteGenTests/AboutExhaustion.cs
@@ -3,30 +3,25 @@ using GalaxyCheck;
 using NebulaCheck;
 using System;
 using System.Linq;
-using Xunit;
 using Property = NebulaCheck.Property;
+using Test = NebulaCheck.Test;
 
 namespace Tests.V2.GenTests.InfiniteGenTests
 {
     public class AboutExhaustion
     {
-        [Fact]
-        public void ItExhaustsWithAnImpossiblePredicate()
-        {
-            NebulaCheck.Property<object> property = new Property(
-                from elementGen in DomainGen.Gen()
-                from seed in DomainGen.Seed()
-                from size in DomainGen.Size()
-                select Property.ForThese(() =>
-                {
-                    var gen = elementGen.Where(_ => false).InfiniteOf();
+        [Property(Iterations = 10)]
+        public NebulaCheck.IGen<Test> ItExhaustsWithAnImpossiblePredicate() =>
+            from elementGen in DomainGen.Gen()
+            from seed in DomainGen.Seed()
+            from size in DomainGen.Size()
+            select Property.ForThese(() =>
+            {
+                var gen = elementGen.Where(_ => false).InfiniteOf();
 
-                    Action test = () => gen.SampleOne(seed: seed, size: size).Take(1).ToList();
+                Action test = () => gen.SampleOne(seed: seed, size: size).Take(1).ToList();
 
-                    test.Should().Throw<GalaxyCheck.Exceptions.GenExhaustionException>();
-                }));
-
-            property.Assert(iterations: 10);
-        }
+                test.Should().Throw<GalaxyCheck.Exceptions.GenExhaustionException>();
+            });
     }
 }
